requirements.txt
selenium
pymongo
python-dotenv